src/requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform

azure-functions
requests==2.32.3
wyze-sdk==2.2.0
slack_sdk==3.27.2
azure-identity==1.16.0
azure-keyvault-secrets==4.8.0
cffi
azure-data-tables==12.5.0
pytz==2024.1
slack-bolt==1.19.0
fastapi==0.111.0
json-log-formatter==1.0